avniproject#762 | StringUtils shifted to S.java
