--- avni-server-api/src/main/java/org/avni/server/util/StringUtils.java
@@ -1,10 +1,0 @@
-package org.avni.server.util;
-
-public class StringUtils {
-    public static String toSnakeCase(String input) {
-        if (input == null) {
-            return null;
-        }
-        return input.trim().replaceAll(" +", "_").toLowerCase();
-    }
-}
